Add a power limit option.
Some NeoPixel setups have limits on how much power they can safely can
draw. For example a NeoPixel feather wing should only draw 1A
sustained, and in practice the power supply of most feathers won't
actually support that high a current.

While we want make the code running on the Arduino 'dumb', having it
protect itself against bugs in the remote program is a good idea.

Empirically the current draw of a NeoPixels is linear in the sum of
the 3 (or 4) bytes that it has been set too. Setting the POWER_LIMIT
macro will cause neoudp.ino to limit the sum of bytes sent to
NeoPixels to POWER_LIMIT. If the sum of the byte values in the pixel
data is greater than POWER_LIMIT then each value will be scaled down
by sum / POWER_LIMIT.

Also fix up README.md to use NeoPixel in place of neopixel and when
iterating over the pixel data use a pointer that starts at the pixel
data instead of constantly having to add an offset.