Accommodate fact that SO_REUSEADDR doesn't replicate received packets to all sockets.
Previous scheme for discovery assumed that with SO_REUSEADDR all the
sockets listing on a given UDP port would get a copy of each received
packet. Turns out this isn't true (or true on all platforms?).

Since we want to be able to run multiple neoudp clients on the same
machine switch DiscoveryListener (nee Listener) to open a socket on an
unused ephemeral port. As part of this narrow the scope of Listener to
just dealing with the discover protocol (which was true before code
wise, but the name and comments implied more).

Also add convenience function that provides a basic
discovery-failure-retry loop for simple clients.